+++ layouts/PricingPageLayout.tsx
@@ -1,9 +1,5 @@
-import { Fragment } from "react"
-
-import ContactForm from "@/components/ContactForm"
 import Image from "@/components/CustomImage"
 import HoverBulletNavMenuAndGetConnectedButton from "@/components/HoverBulletNavMenuAndGetConnectedButton"
-import IconCard from "@/components/IconCard"
 import IconColumnSection from "@/components/IconColumnSection"
 import OurTeamSection from "@/components/OurTeamSection"
 import PricingPackagesBlueSection from "@/components/PricingPackagesBlueSection"
@@ -12,7 +8,8 @@ import PricingPackagesSectionFootnotesAndDetails from "@/components/PricingPacka
 import { PageSeo } from "@/components/SEO"
 import TitleHeadingsAndChildren from "@/components/TitleHeadingsAndChildren"
 import siteMetadata from "@/data/siteMetadata.json"
-import { classNames } from "@/lib/utils"
+
+import { ContactFormSection } from "../components/ContactFormSection"
 
 export default function PricingPageLayout({
   slug,
@@ -86,52 +83,5 @@ export default function PricingPageLayout({
         />
       </div>
     </>
-  )
-}
-
-function ContactFormSection({
-  officeAddressLeft,
-  officeAddressRight,
-  contactForm,
-}: {
-  officeAddressLeft?: IconCard[]
-  officeAddressRight?: IconCard[]
-  contactForm?: ContactField[]
-}) {
-  // the contact form is always required
-  if (!contactForm) return null
-  // if either is present, both left and right office addresses are required
-  if (officeAddressLeft && !officeAddressRight) return null
-  if (!officeAddressLeft && officeAddressRight) return null
-  return (
-    <div
-      className={classNames(
-        "grid max-w-5xl mx-auto",
-        officeAddressLeft && officeAddressRight
-          ? "grid-cols-1 md:grid-cols-3 auto-cols-min gap-6 px-2"
-          : ""
-        // 3 column layout is used for /contact and 1 column for /order
-      )}
-    >
-      {officeAddressLeft && (
-        <OfficeAddressColumn iconCards={officeAddressLeft} />
-      )}
-      {contactForm && <ContactForm contactForm={contactForm} />}
-      {officeAddressRight && (
-        <OfficeAddressColumn iconCards={officeAddressRight} />
-      )}
-    </div>
-  )
-}
-
-function OfficeAddressColumn({ iconCards }: { iconCards: IconCard[] }) {
-  return (
-    <div>
-      {iconCards.map((iconCard: IconCard) => (
-        <Fragment key={`${iconCard.text}${iconCard.label}`}>
-          <IconCard {...iconCard} />
-        </Fragment>
-      ))}
-    </div>
   )
 }
